requirements.txt
# Slack
slack-bolt>=1.23.0
slack-sdk>=3.36.0

# Google AI
google-generativeai>=0.3.0

# Vector Database
chromadb>=0.4.0

# Environment Variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.5.0